refactor(app): clean new technical
