+++ src/components/atoms/__tests__/button.test.tsx
@@ -2,9 +2,9 @@ import { render } from '@testing-library/react';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
 import { Smiley } from '@phosphor-icons/react';
-import { Button } from '.';
+import { Button } from '../button';
 
-describe('components button', () => {
+describe('components/atoms button', () => {
   const Icon = Smiley;
 
   afterEach(() => {
