+++ src/components/atoms/__tests__/no-ssr.test.tsx
@@ -1,9 +1,9 @@
 import { render } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 
-import { NoSsr } from '.';
+import { NoSsr } from '../no-ssr';
 
-describe('components no-ssr', () => {
+describe('components/atoms no-ssr', () => {
   it('renders correctly', () => {
     const result = render(<NoSsr>Click me</NoSsr>);
