--- src/types/index.tsx
@@ -1,2 +1,0 @@
-export * from './media';
-export * from './theme';
